backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
huggingface_hub
python-dotenv==1.0.1
pydantic==2.10.6
